+++ CHANGES.md
@@ -19,11 +19,13 @@
     -   Currently uses only a case-sensitive check, using 'in' (i.e. if the filtering parameter is inside the job-name, then it passes the check).
 -   Added `--sort` argument to sort 'sacct' results by the given field.
     -   New method `sacct._sort_lines()` does this, called from `sacct.sacct_results()`.
+-   By default, reformat the times-strings output by sacct (replace 'T' separator with optional alternative character.  Set by parameters in `const.py`, with code in `sacct._parse_sacct_line()`.
 -   By default, printed output now only shows a subset of available parameters.  Use '--verbose' to include all of them.
     -   New method `utils._filter_fields_in_lines()` is called from within `utils.print_lines_dicts()`.
 -   `slurpy/sacct.py`
     -   `_construct_sacct_command()`
         -   New internal method to construct the sacct command to be called, allow optional arguments, etc.
+
 
 ## v0.2 - 2017/01/14
 ## -----------------
@@ -49,6 +51,7 @@
         -   Method for pretty-printing lines retrieved from 'sacct'.
     -   `prompt_yes_no` [new-method]
         -   Prompt the user for a y/n answer for an arbitrary question, take5n from [Recipe 57708](http://code.activestate.com/recipes/577058/).
+
 
 ## v0.1 - 2016/12/13
 ## -----------------